Remove old netlist API and IRODESystem
This commit removes the old DAECompiler-based netlist API and IRODESystem:

**Deleted files:**
- src/circuitodesystem.jl - CircuitIRODESystem, DefaultSim, ParamSim
- src/mna/stubs.jl - DAECompiler stubs (IRODESystem, variable, equation, etc.)
- src/deprecated.jl - deprecated conversion functions
- src/dcop.jl - CedarDCOp, CedarTranOp initialization algorithms
- src/simulate_ir.jl - Net, branch!, DAECompiler primitives
- src/simpledevices.jl - SimpleResistor, VoltageSource, etc. (DAECompiler-based)
- ext/CedarSimCSVExt.jl - CSV extension (depended on CircuitIRODESystem)
- ext/CedarSimPlotlyLightExt.jl - PlotlyLight extension (same)

**Modified files:**
- src/CedarSim.jl - Removed USE_DAECOMPILER, added minimal stubs for
  backward compatibility (ParallelInstances, CircuitElement, etc.)
- src/spectre_env.jl - Removed DAECompiler conditionals and device aliases
- src/va_env.jl - Removed DAECompiler imports, stubbed ddt/noise functions
- src/util.jl - Removed DAECompiler ddt import
- src/sweeps.jl - Removed DAECompiler arg1_from_sys import
- src/netlist_utils.jl - Simplified, removed DAECompiler dependencies
- src/circsummary.jl - Removed DAECompiler method extensions
- Project.toml - Removed CSV and PlotlyLight extensions
- test/common.jl - Removed DAECompiler-based solve functions
- test/spectre_expr.jl - Removed USE_DAECOMPILER check

**Kept for reference (commented out):**
- src/ac.jl - AC/noise analysis (include commented out)
- src/aliasextract.jl - alias extraction (include commented out)

The MNA backend (src/mna/) is now the sole simulation engine.
563 MNA tests pass.
