+++ src/sweeps.jl
@@ -3,13 +3,6 @@ using AxisKeys
 using OrdinaryDiffEq, SciMLBase, Sundials
 using Base.Iterators
 
-# Phase 0: Use stubs instead of DAECompiler
-@static if CedarSim.USE_DAECOMPILER
-    using DAECompiler: arg1_from_sys
-else
-    using ..DAECompilerStubs: arg1_from_sys
-end
-
 export alter, dc!, tran!, Sweep, CircuitSweep, ProductSweep, TandemSweep, SerialSweep, sweepvars, split_axes, sweepify
 
 # This `alter()` is to make it easy to apply directly to a struct or named tuple with the optics.
